+++ src/main/java/com/hoshimusubi/suhwa/dto/CommentsDTO.java
@@ -10,9 +10,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class CommentsDTO {
-	private Long id;
-    private Long postId;              // 댓글이 속한 게시글 ID
-    private Long userId;              // 작성자 ID
+	private int id;
+    private int postId;              // 댓글이 속한 게시글 ID
+    private int userId;              // 작성자 ID
     private String content;
     private LocalDateTime created_at;
     private String nickname;
