+++ src/main/java/com/hoshimusubi/suhwa/dto/LikesDTO.java
@@ -11,8 +11,8 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class LikesDTO {
-	private Long id;
-    private Long userId;
-    private Long postId;
+	private int id;
+    private int userId;
+    private int postId;
     private LocalDateTime createdAt;
 }
