+++ src/main/java/com/hoshimusubi/suhwa/dto/MessageDTO.java
@@ -10,9 +10,9 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MessageDTO {
-	private Long id;
-    private Long senderId;
-    private Long receiverId;
+	private int id;
+    private int senderId;
+    private int receiverId;
     private String content;
     private Integer isRead;             // 0 or 1
     private LocalDateTime sentAt;
