+++ src/main/java/com/hoshimusubi/suhwa/mapper/MessageMapper.java
@@ -12,6 +12,6 @@ public interface MessageMapper {
 
 	void insertMessage(MessageDTO message);
 
-	UsersDTO findUserById(Long userId);
+	UsersDTO findUserById(int userId);
 
 }
