+++ src/main/java/com/hoshimusubi/suhwa/service/BookMarkService.java
@@ -5,8 +5,8 @@ import org.apache.ibatis.annotations.Param;
 import com.hoshimusubi.suhwa.dto.UsersDTO;
 
 public interface BookMarkService {
-	void addBookmark(Long postId, @Param("userId")Long loginUser);
-    void removeBookmark(Long postId, @Param("userId")Long userId);
-    boolean isPostBookmarkedByUser(Long postId, @Param("userId")Long long1);
+	void addBookmark(int postId, @Param("userId")int loginUser);
+    void removeBookmark(int postId, @Param("userId")int userId);
+    boolean isPostBookmarkedByUser(int postId, @Param("userId")int long1);
 
 }
