+++ src/main/java/com/hoshimusubi/suhwa/service/MessageService.java
@@ -7,6 +7,6 @@ public interface MessageService {
 
 	void sendMessage(MessageDTO message);
 
-	UsersDTO getUserById(Long userId);
+	UsersDTO getUserById(int userId);
 
 }
