+++ src/main/java/com/hoshimusubi/suhwa/service/PostsService.java
@@ -3,9 +3,9 @@ package com.hoshimusubi.suhwa.service;
 import com.hoshimusubi.suhwa.dto.PostsDTO;
 
 public interface PostsService {
-	PostsDTO getPostById(Long id);
+	PostsDTO getPostById(int id);
 	void savePost(PostsDTO post);
     void updatePost(PostsDTO post);
-    void deletePost(Long id);
-    void increaseViewCount(Long postId);
+    void deletePost(int id);
+    void increaseViewCount(int postId);
 }
